requirements.txt
requests>=2.28.0
discord.py>=2.0.0